Communication/requirements.txt
SpeechRecognition==3.10.0
PyAudio==0.2.11
murf
requests==2.31.0
python-dotenv==1.0.0
pygame==2.6.1 